Added system where if a user does not use his gotten code within a week (the user did not go to the event), the user loses points.

+++ src/main/java/be/thomasmore/setalight/models/Profile.java
@@ -23,7 +23,7 @@ public class Profile {
     private String physicalAttributes;
     private String nationalInsuranceNumber;
     private double length;
-    private int rewardPoints;
+    private double rewardPoints;
     @ManyToMany
     private Collection<Profile> friends;
     private String height;
@@ -151,11 +151,11 @@ public class Profile {
         this.checkedEvents = checkedEvents;
     }
 
-    public int getRewardPoints() {
+    public double getRewardPoints() {
         return rewardPoints;
     }
 
-    public void setRewardPoints(int rewardPoints) {
+    public void setRewardPoints(double rewardPoints) {
         this.rewardPoints = rewardPoints;
     }
